+++ src/main/kotlin/hilfling/backend/hilfling/controller/PhotoGangBangerController.kt
@@ -1,11 +1,7 @@
 package hilfling.backend.hilfling.controller
 
-import hilfling.backend.hilfling.model.Album
 import hilfling.backend.hilfling.model.PhotoGangBanger
-import hilfling.backend.hilfling.repository.AlbumRepository
 import hilfling.backend.hilfling.repository.PhotoGangBangerRepository
-import me.liuwj.ktorm.database.Database
-import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.web.bind.annotation.*
 
 @RestController
@@ -14,7 +10,7 @@ class PhotoGangBangerController {
     val repository = PhotoGangBangerRepository()
 
     @GetMapping("/{id}")
-    fun getById(@PathVariable("id") id: Long): PhotoGangBanger? {
+    fun getById(@PathVariable("id") id: Int): PhotoGangBanger? {
         return repository.findById(id)
     }
