+++ src/main/kotlin/hilfling/backend/hilfling/controller/PurchaseOrderController.kt
@@ -1,11 +1,7 @@
 package hilfling.backend.hilfling.controller
 
-import hilfling.backend.hilfling.model.Album
 import hilfling.backend.hilfling.model.PurchaseOrder
-import hilfling.backend.hilfling.repository.AlbumRepository
 import hilfling.backend.hilfling.repository.PurchaseOrderRepository
-import me.liuwj.ktorm.database.Database
-import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.web.bind.annotation.*
 
 @RestController
@@ -14,7 +10,7 @@ class PurchaseOrderController {
     val repository = PurchaseOrderRepository()
 
     @GetMapping("/{id}")
-    fun getById(@PathVariable("id") id: Long) : PurchaseOrder? {
+    fun getById(@PathVariable("id") id: Int) : PurchaseOrder? {
         return repository.findById(id)
     }
